Add informing about each macro use
Lexer listener has been extended with a method that is called when a macro
use in the source code is detected. Information about that use of a macro is
saved and passed in an object of a new token class, MacroToken. This object
will usually carry location of the macro definition.