Handle unexpected interface reference
When we try to refer to the typedef name defined in a component specification
section, and the typedef does not exist, due to the context sensitiveness
issues, the interface reference construct is recognized. Instead of throwing
an exception, return an error message.

However, the context sensitiveness is still not resolved.