Remove collecting tokens
The first approach to the plug-in required the frontend to collect detailed
data of each token: keywords, punctuators, etc. Since these tokens are
no longer needed, we can stop collecting them. However, the collection
mechanism should stay, because it may be used for indicating different
types of identifiers: variables, function names etc. This distinction may
help the plug-in to quickly determine the token pointed by the user in
the editor, and to perform some action, e.g. jumping to the declaration,
finding usages etc.