Enhance the analysis of tags
The structure of fields of a field tag type is now built and stored in the
symbol table. It is useful if a tag contains an unnamed member of an anonymous
tag type. Moreover, types of fields of field tag types are checked for
validity and detected errors are reported.