Add basic type control of variables declarations
It is checked if variables with no linkage are of complete types. If not,
errors are reported. Moreover, non-type specifiers from declarations of
variables are analyzed and if a bad combination is detected, an error helper
is notified about such error. Furthermore, the builder pattern is introduced
to the hierarchy of classes that depict objects in the symbol table. Now the
only way to instantiate objects of these classes is to use a builder.