Create class hierarchy for specific issues
A class hierarchy that reflects errors and warnings about analyzed programs is
designed and implemented. It is located in package:

pl.edu.mimuw.nesc.problem.issue

Currently only issues that are related to types resolving are included in this
hierarchy - the code responsible for that process has been changed and its part
that was generating issues messages has been removed. Moreover, a package for
AST utilities is created and appropriate classes are moved to it.