Adapt the parser to different kinds of literals
AST nodes for integer, character and floating literals are created. The parser
now distinguishes them and makes appropriate nodes that represent them. Integer
literals are analyzed the most before creating nodes - their kinds and suffixes
are determined and stored in nodes in the form of enum values. Numerical values
of the literals are also determined and stored if they don't exceed the maximum
integer value that is allowed. Moreover, tests are created to check the
correctness of analysis of these literals.