Refactor codes of issues
Every issue type is now associated with a unique and only instance of Code.
This instance is always returned by 'getCode()' method in the appropriate class.
It allows to uniquely identify an issue that has appeared. Moreover,
'NescIssue' class now carries the code of the issue as an optional value. If an
error or warning is reported using an object of the issues hierarchy, then the
code is stored in the object of a class derived from 'NescIssue'.