Add integration tests
Integration tests are implemented. They consist of good and bad tests. Good
tests are associated with correct programs and the compiler must not report any
errors and warnings for them. In turn, bad tests are associated with programs
with exactly one issue that must be detected by the compiler. Moreover, basic
test programs for the tests are added.