Refactor tags analysis
The method of tags analysis is changed significantly. Currently there is
a bijection between created tag declaration objects and different tag types
(except that there are no objects for invalid tag types). Previously, there
was a bijection between AST nodes of tag references and tag declaration
objects (object for a new tag reference overwrote the previous object
if present in the symbol table). Now the tag declaration object is updated
when necessary and not overwritten.

Moreover, the builder pattern is introduced to the tag declarations hierarchy,
contained in package:

pl.edu.mimuw.nesc.declaration.tag

In turn, the factory pattern is used for creating appropriate objects for
abstract tag references during the analysis of tags.