Implement a facade hierarchy for interface references
A small hierarchy of facade classes for interface references is created. It is
intended to provide an easy access to the information about the reference and
the interface itself, e.g. about commands and events declared in the interface.
The substitution of types if the interface is generic is handled transparently
by a facade. The facade for an interface reference is available after calling
method:

pl.edu.mimuw.nesc.declaration.object.InterfaceRefDeclaration.getFacade()

after the whole specification of the component has been parsed and processed.