Add checking completeness of modules implementations
It is checked if all provided commands and events of a module are implemented.
If not, an error is reported for each missing definition. Moreover, code
responsible for analysis of functions definitions is extended. It is checked if
prototypes of implemented commands and events match their declarations in
interfaces or modules specifications. Tasks definitions are also checked.

The concept of a module table is introduced. It is a kind of a symbol table for
a module - it contains information about all commands and events that a module
must (provided commands and events) or can (used commands and events)
implement. It is located in a ModuleDeclaration object and contains full
information after parsing and analysis of the whole module.