Mark AST nodes that refer to generic parameters
Two AST nodes classes: 'Typename' and 'Identifier' are enriched with a flag
that is raised if the node refers to a generic parameter of the component it
occurs in (if it does not occur in any component the flag is lowered).
Moreover, expressions that appear in array declarators, initializers and
enumerators definitions are checked.