Enhance support for NesC constant functions
New AST nodes 'UniqueCall', 'UniqueNCall' and 'UniqueCountCall' are added. They
represent calls to NesC compile-time constant functions and are created when
such calls appear in the program (instead of 'FunctionCall' nodes). Moreover,
correctness of the calls is checked. For instance, if the first provided
parameter is not a constant string, an error is reported.