Implement evaluation of constant functions
Evaluation of compile-time NesC constant functions ('unique', 'uniqueN' and
'uniqueCount') is implemented. Objects responsible for it store computed values
in AST nodes of calls to these functions. Moreover, the substitution of generic
parameters is modified not to create new AST nodes of calls to the constant
functions. It causes that every usage of 'unique', 'uniqueN' or 'uniqueCount'
will be evaluated exactly once.