Add handling of @C() attribute
It is checked if the usage of @C() attribute is correct. If not, an error is
reported. Moreover, names of all global entities and entities marked by @C()
attribute are collected for the operation of finalizing names mangling. A map
with gathered names from the entire application is accessible by
a 'ProjectData' object.