Implement loading and checking the scheduler component
A new option '--scheduler' is added. It specifies the scheduler to be used by
a NesC application. A proper scheduler component is loaded by the frontend and
then check for correctness. All errors detected in the scheduler component and
in the task interface are added to project issues of the NesC application.