Implement tasks transformation
The transformation of tasks to interface references and wiring with the
scheduler component is implemented. A new task interface reference is created
for each task. Moreover, tasks implementations are changed to events
implementations and posting expressions to command calls to the scheduler
component. Furthermore, the created interface references are wired to the
parameterised interface provided by the scheduler component.