Add collecting information about combining functions
Project data contains now information about all combining functions declared
in the entire project. Moreover, usages of '@combine' attribute are checked
for correctness and detected errors are reported. Furthermore, analysis of
attributes is refactored - it is now done by an object of class:

'pl.edu.mimuw.nesc.analysis.AttributeAnalyzer'

The object checks the attributes and performs actions that they imply using
a chain of small analyzers. Each of the small analyzers is responsible for
a single type of attributes, e.g. '@C()'.