Collect unique names of commands, events and tasks
The module table is enriched with a method that allows making its deep copy.
Moreover, it allows storing unique names of commands, events and tasks it
contains. Furthermore, modules tables are copied for instantiated generic
modules. After instantiation of a generic module remangled unique names of
commands, events and tasks are collected and stored in the copied module
table.