Add creating empty intermediate functions
Each node in the wires graph contains now a data object with information about
the corresponding function. For commands and events provided by modules, it is
only the unique name of the implementation of the command or event. For
commands and events that are not provided or are specified in configurations,
it is the AST node of the empty intermediate function along with immutable list
with names of its parameters and other information that facilitates
constructing intermediate functions.