Add writing AST nodes as source code
A class that may be used to write source code of a NesC or C program
represented by an AST node is implemented:
   'pl.edu.mimuw.nesc.astwriting.ASTWriter'
All AST nodes can be written (except some that are not used, e.g. Conjugate).
Currently basic formatting of written code is provided. Moreover, the structure
of packages in 'nesc-ast' module is changed - packages:
   * 'pl.edu.mimuw.nesc.ast.util'
   * 'pl.edu.mimuw.nesc.ast.type'
are moved to package 'pl.edu.mimuw.nesc' and the first is slightly renamed.