Copy objects from the declaration hierarchy during instantiation
Objects from the declaration hierarchy ('pl.edu.mimuw.nesc.declaration') that
represent functions, variables, constants, typenames and other entities are
copied during the instantiation of components for each created component. Their
contents are updated to be valid for the new component. The declaration
hierarchy is equipped with a mechanism for copying its objects. The most
important elements of this mechanism:
   * class 'pl.edu.mimuw.nesc.declaration.CopyController'
   * method 'pl.edu.mimuw.nesc.declaration.Declaration#deepCopy' that must be
     implemented in each declaration object class that is not abstract