Add checking definitions of external structures and unions
It is checked during the analysis of tags if external structures and external
unions contain only members of external types. If such structure or union
contains a member of type that is definitely not external, then an error is
reported. For fields that may be external after instantiation of components no
error is reported but they need to be checked again after the instantiation.