Prepare tag declarations classes for computing types layout
Declaration objects that represent tags can now contain the size and alignment
of the corresponding tag types. Moreover, field declarations objects can hold
the offset, size and alignment of the fields they represent. Furthermore, the
old concept of storing all information necessary for computing the size and
alignment of a type directly in a 'Type' object is abandoned.