Add hierarchy of constant values
The classes hierarchy that represent constant values is implemented. It
represent values that will be returned by the interpreter of constant
expressions. The hierarchy will be extended with new operations such
as multiplication or division. It contains constant types hierarchy that
allows type-safe evaluation of constant expressions and provides information
necessary for performing operations on constants. Moreover, the declaration
hierarchy is improved, e.g. the objects that represent enumeration constants
allow to store their values.