Implement traversing intermediate functions generator
A new intermediate functions generator is implemented and used in the
default backend. The generator resolves all connections to implementations
at compile-time. Intermediate functions that correspond to used commands or
events from modules are the only intermediate functions that are generated.
All of them call implementations of corresponding commands and events
directly. Result intermediate functions are no longer necessary and aren't
generated.