Combine main and type declarations cleaners
Main and type declarations cleaners are combined to one cleaner class. It is
necessary for correctness of cleaning variable and function declarations
because, for example, a global variable may be only referred in a sizeof
operator in a bit-field width expression and in such case it cannot be
removed (in the approach of two separate cleaners it would).