Disable tasks optimization if it can be dangerous
Tasks optimization is disabled when the unique identifier used for connecting
tasks to the scheduler is used elsewhere or if some tasks are connected without
usage of 'unique' with the identifier. There are also some other cases when the
optimization is disabled. A warning is printed if the optimization is requested
by a compiler option but becomes disabled.