Extract the code of default compilation
The code responsible for the default compilation that transforms a NesC program
to a list of C declarations and definitions in proper order is extracted to
a new package 'pl.edu.mimuw.nesc.compilation'. It is used now in the default,
GCC backend and will be used in the SDCC/8051 microcontrollers backend.