Implement simple code partitioner
A simple code partitioner that partitions functions into code banks is
implemented. The algorithm is as follows. Spontaneous functions are
allocated to the common bank. All other functions are assigned in the
descending order of their sizes: a function is assigned to the bank with
the least amount of remaining space that can hold it.