Add rigid functions option to 8051 backend
An option that allows specifying a list of rigid functions is added to the
8051 backend. A rigid function is a function whose banking characteristic
is not changed by the compiler. It means that if a rigid function is
declared as banked, it remains banked in the output C files and if it is
not declared as banked, it remains not banked in the output files.