Change concept for handling banked characteristic
The 8051 backend behaves now differently when it encounters a function marked
as banked. A new option '--relax-banked' allows controlling the behaviour. If
this option is not given, then all functions marked as banked will be banked
in the output files. However, if this option is specified, then it is possible
that a defined function that is not spontaneous will become not banked in the
output files. The concept of rigid functions is abandoned and the option for it
('--rigid-functions') is removed.