Generalize the model of banks
The model of banks is generalized from positive number of banks of fixed size
with the common bank to positive number of banks that can differ in size with
the common bank. Now each bank can have an arbitrary non-negative size and an
arbitrary name which can be specified using a new option '--banks'. Options
'--banks-count' and '--bank-size' are removed.

Moreover, new abstractions for the code partition operation are developed: bank
schema and bank table. A bank schema carries information about available banks
(their names, sizes and name of the common bank) and a bank table allows
allocating functions to banks defined in a bank schema.