Check if used commands and events are implemented
It is checked for every call of a used command and every signal of a used
event if the command or the event is implemented, i.e. if it is connected
to an implementation or a default implementation is provided for it. If it
is not the case, then a compilation error is reported and the compilation
is terminated.