Correct handling errors in programs
When an interface reference declarator is used in a declaration that is not
the definition of a function, a declaration object is created and associated
with the declaration. This prevents for throwing an exception in further
analysis. A definition of a command or an event can be perceived by the
compiler as a declaration that is not a definition when an undefined or
shadowed typedef is used as the return type of the command or event.