Enrich the biconnected components heuristic
The biconnected components heuristic is enriched with the following features.
First, it operates on the maximum spanning forest of the call graph instead
of the spanning forest constructed when computing biconnected components.
Weights on edges used for computation of the maximum spanning forest are their
frequency estimations. Second, for each candidate allocation it computes the
sum of frequency estimations of all functions in the allocation. The sum is
a new criterion used for choosing the best allocation: an allocation with
a greater frequency estimations sum is preferred. All previously used
criterions are applied when the new criterion does not indicate a better
allocation.