Add an option for choosing a spanning forest kind
A new option, '--spanning-forest', is added to the 8051 backend. It allows for
choosing the kind of the spanning forest that will be used by the biconnected
components heuristic for the partitioning. A minimum and maximum spanning
forest can be selected. Moreover, the spanning forest constructed during the
computation of biconnected components can be used.