Add an option for preferring higher estimate allocations
A new option, '--prefer-higher-estimate-allocations' is added to the 8051
backend. It turns on the change of the criterion used by the biconnected
components heuristic to select an allocation to made, introduced by commit
f6287d6563c252554ec9a5f12cc1ebd1fc480f3f. If this option is not specified,
then the criterion for choosing the allocation is as if the aforementioned
change was not made.