Implement extended subtree partitioning
If option '--extended-subtree-partitioning' is specified for the 8051 backend,
operations performed by the biconnected components are different: if there is
no allocation of an entire tree or in a cut vertex that fits a bank, down- and
up-tree allocations in all vertices are considered. The allocation that
minimizes the sum of weights of edges incident to a vertex in the allocation
and to a vertex outside the allocation is chosen and made. If there are
multiple such allocations, then the best one is selected amongst them using the
criterions for normal candidate allocations (in cut vertices).