Improve extended subtree partitioning
Extended subtree partitioning is renamed to arbitrary subtree partitioning.
Option '--extended-subtree-partitioning' is replaced with option
'--arbitrary-subtree-partitioning' that allows for specifying when the
arbitrary subtree partitioning is performed. The content of the common
bank and the target bank of an allocation are now taken into consideration
when the quality of an allocation is estimated during the arbitrary subtree
partitioning.