Implement new methods for picking allocations
New methods for picking allocations can be selected for the arbitrary subtree
partitioning in the biconnected components heuristic. A method can be chosen
by using option '--allocation-picker-asp'. An allocation with the greatest
functions size amongst a specific number of allocation with the smallest
weights sums or an allocation with the smallest weights sum amongst a specific
number of allocations with the greatest functions sizes can be chosen.